+++ lld/ELF/Symbols.cpp
@@ -541,11 +541,8 @@ void elf::reportDuplicate(Ctx &ctx, const Symbol &sym, const InputFile *newFile,
                           InputSectionBase *errSec, uint64_t errOffset) {
   if (ctx.arg.allowMultipleDefinition)
     return;
-  // In glibc<2.32, crti.o has .gnu.linkonce.t.__x86.get_pc_thunk.bx, which
-  // is sort of proto-comdat. There is actually no duplicate if we have
-  // full support for .gnu.linkonce.
   const Defined *d = dyn_cast<Defined>(&sym);
-  if (!d || d->getName() == "__x86.get_pc_thunk.bx")
+  if (!d)
     return;
   // Allow absolute symbols with the same value for GNU ld compatibility.
   if (!d->section && !errSec && errOffset && d->value == errOffset)